Sort buffers generate the same result even thought the option autochdir was set.
When the option autochdir was set, the path to the buffers is relative to the CWD.
When the activity buffer was changed, the CWD was changed. And then, the buffer list
was different after sort. And then, some buffer will never change to, when use
CtrlSpaceGoDown or CtrlSpaceGoUp command.
This commit will use the absolute path to sort buffer. The bug will be fix.

This a way to reply the bug:
0. `set autochdir` in ~/.vimrc
1. cd to '~/.vim/bundle/vim-ctrlspace/autoload/ctrlspace'
2. open api.vim with vim
3. use ':e search.vim', ':e keys/buffer.vim' to open them
4. the ':CtrlSpaceGoDown' command, will jump betweem  api.vim and buffer.vim,
    the search.vim will never be opened.